Factor out common uninstall code
