+++ components/hab/src/command/pkg.rs
@@ -19,6 +19,7 @@ pub mod provides;
 pub mod search;
 pub mod sign;
 pub mod uninstall;
+pub mod uninstall_impl;
 pub mod upload;
 pub mod verify;
